+++ src/test/java/com/unzer/payment/business/CancelAfterAuthorizationTest.java
@@ -29,6 +29,7 @@ import java.math.BigDecimal;
 import java.net.MalformedURLException;
 
 import static com.unzer.payment.business.BasketV1TestData.getMaxTestBasketV1;
+import static com.unzer.payment.business.Keys.MARKETPLACE_KEY;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -144,18 +145,18 @@ public class CancelAfterAuthorizationTest extends AbstractPaymentTest {
             maxBasket.getBasketItems().get(i).setAmountDiscount(null);
         }
 
-        Basket basket = getUnzer(marketplacePrivatekey).createBasket(maxBasket);
+        Basket basket = getUnzer(MARKETPLACE_KEY).createBasket(maxBasket);
 
         // create card
         Card card = getPaymentTypeCard(NO_3DS_VISA_CARD_NUMBER); //do not change card number except error case
-        card = getUnzer(marketplacePrivatekey).createPaymentType(card);
+        card = getUnzer(MARKETPLACE_KEY).createPaymentType(card);
 
         // marketplace authorization
         MarketplaceAuthorization authorizeRequest = getMarketplaceAuthorization(card.getId(), null, null, null,
                 basket.getId(), null);
         authorizeRequest.setAmount(maxBasket.getAmountTotalGross());
 
-        MarketplaceAuthorization authorize = getUnzer(marketplacePrivatekey).marketplaceAuthorize(authorizeRequest);
+        MarketplaceAuthorization authorize = getUnzer(MARKETPLACE_KEY).marketplaceAuthorize(authorizeRequest);
         assertNotNull(authorize.getId());
         assertNotNull(authorize);
         assertEquals(AbstractTransaction.Status.PENDING, authorize.getStatus());
@@ -192,18 +193,18 @@ public class CancelAfterAuthorizationTest extends AbstractPaymentTest {
             maxBasket.getBasketItems().get(i).setAmountDiscount(null);
         }
 
-        Basket basket = getUnzer(marketplacePrivatekey).createBasket(maxBasket);
+        Basket basket = getUnzer(MARKETPLACE_KEY).createBasket(maxBasket);
 
         // create card
         Card card = getPaymentTypeCard(NO_3DS_VISA_CARD_NUMBER); //do not change card number except error case
-        card = getUnzer(marketplacePrivatekey).createPaymentType(card);
+        card = getUnzer(MARKETPLACE_KEY).createPaymentType(card);
 
         // marketplace authorization
         MarketplaceAuthorization authorizeRequest = getMarketplaceAuthorization(card.getId(), null, null, null,
                 basket.getId(), null);
         authorizeRequest.setAmount(maxBasket.getAmountTotalGross());
 
-        MarketplaceAuthorization authorize = getUnzer(marketplacePrivatekey).marketplaceAuthorize(authorizeRequest);
+        MarketplaceAuthorization authorize = getUnzer(MARKETPLACE_KEY).marketplaceAuthorize(authorizeRequest);
         assertNotNull(authorize.getId());
         assertNotNull(authorize);
         assertEquals(AbstractTransaction.Status.PENDING, authorize.getStatus());
@@ -215,7 +216,7 @@ public class CancelAfterAuthorizationTest extends AbstractPaymentTest {
         assertEquals(HttpStatus.SC_MOVED_TEMPORARILY, redirectStatus);
 
         //fetch payment
-        MarketplacePayment payment = getUnzer(marketplacePrivatekey).fetchMarketplacePayment(authorize.getPaymentId());
+        MarketplacePayment payment = getUnzer(MARKETPLACE_KEY).fetchMarketplacePayment(authorize.getPaymentId());
 
         //partial cancel
         MarketplaceCancel cancelRequest = new MarketplaceCancel();